TextboxList: add mustMatch option and make it accessible in DataForms
In certain use cases, we want to discourage users from entering a
completely random value into a field. This introduces a 'mustMatch'
option that will only accept a value if it came from the autocomplete
list. Invalid values will be removed and the input box will flash.

To make this accessible from DataForms, also introduce a mechanism to
pass options to TextboxList from within a form definition. Ideally the
options would be added elsewhere, but the better candidates didn't work
out:

   * the 'Attributes' column looks like an ideal fit but the current FW
     code determines mandatoriness of a field by matching the attributes
     against /M/. Passing options in that field is bound to confuse the
     match.
   * the 'Type' column would be a close second but the current FW code
     downcases the field, so not all option names (and values) could be
     used here.

So, as a somewhat ugly alternative, parse the options out of the 'Size'
column. They don't really fit there but this should be a column where
they won't get in the way. After they have been parsed out, the 'size'
member of the object is cleaned up, so everyone else gets the size and
only the size.