TextboxList: add useHidden option to collect all values in one field
If 'useHidden' is used, the values will be kept in a comma-separated
form in a single hidden input field; no other hidden input fields will
be created and the 'name' attribute will be removed from the original
input field. That way, when the form is submitted, it will look to
handlers as if the user simply entered a comma-separated list of values
into the field.

This is required for mandatory textboxlist fields in a DataForm to work
correctly.