textboxlist: fix issues with leading/trailing spaces in value list
If the list of values in a textboxlist has leading whitespace in the
first value or trailing whitespace in the last value, the code in core
that attempts to preserve ordering of values from the form can no longer
match up these values and drops them.

Make textboxlist strip whitespace the same way core does it.

Ticket: #10318
Release-Note(fix,user) {
  Resolve a problem due to which values entered with leading or trailing
  spaces could get lost in textboxlist fields (e.g. glossary keywords).
}